ref: move PoguesModelUtils
