+++ src/main/java/fr/insee/pogues/transforms/visualize/composition/InsertSequences.java
@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
 
-import static fr.insee.pogues.utils.PoguesModelUtils.getSequences;
+import static fr.insee.pogues.utils.model.PoguesModelUtils.getSequences;
 
 /**
  * Implementation of CompositionStep to replace questionnaire reference by its sequences.
