+++ src/test/java/fr/insee/pogues/utils/PoguesModelUtilsTest.java
@@ -5,6 +5,7 @@ import fr.insee.pogues.exception.IllegalIterationException;
 import fr.insee.pogues.model.DynamicIterationType;
 import fr.insee.pogues.model.FlowControlType;
 import fr.insee.pogues.model.IterationType;
+import fr.insee.pogues.utils.model.PoguesModelUtils;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
